Add ability to customize tostring() implementation using methamethods on default delegate.
Original ToString implementation does not take default delegates into account. By exposing default to the Squirrel code it should be possible to customize the default string representations of objects (make Tables display their slots for example etc.). I've added invocation of the default delegate's _tostring methamethod when present so this customization is now possible.

Other thing to consider is the recursion check. If a recursive methamethod call is detected, the raw string conversion is applied (i.e. the one which does not invoke any delegates), so it is possible to call the `super` (raw) string conversion by concatenating `this` with a certain sting for example.