Fix nullptr access warning.
This is probably a false positive, but just to make the Clang static analyzer happy, I've added the nullptr check here.

Analyzer constructed a rather lengthy code execution path where it could end up having null _freelist and thus crash on newnode->obj = obj. I'm not really sure that this kind of fix is a good idea, but I haven't actually encountered Squirrel crashes in this place before, so I generally tend to think that this is a false positive and we should not worry about the fix too much.